+++ aoc-19/src/day_02.rs
@@ -12,11 +12,11 @@ impl Fro for ProgramAlarm {
 }
 
 impl Solution for ProgramAlarm {
-    fn one(mut self) -> i32 {
+    fn one(mut self) -> i64 {
         self.0.run_nv(12, 2)
     }
 
-    fn two(mut self) -> i32 {
+    fn two(mut self) -> i64 {
         for noun in 0..100 {
             for verb in 0..100 {
                 self.0.reset();
